requirements.txt
asgiref==3.7.2
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cleo==2.1.0
crashtest==0.4.1
cryptography==44.0.0
distlib==0.3.9
Django==5.0.1
djangorestframework==3.14.0
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.41
identify==2.6.4
idna==3.10
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
mysqlclient==2.2.1
nodeenv==1.9.1
packaging==24.2
pexpect==4.9.0
pkginfo==1.12.0
platformdirs==4.3.6
poetry-core==1.9.1
poetry-plugin-export==1.8.0
pre_commit==4.0.1
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.2.0
pytz==2023.3.post1
PyYAML==6.0.2
RapidFuzz==3.11.0
requests==2.32.3
requests-toolbelt==1.0.0
ruff==0.8.4
SecretStorage==3.3.3
setuptools==69.0.3
shellingham==1.5.4
smmap==5.0.1
sqlparse==0.4.4
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
urllib3==2.3.0
virtualenv==20.28.0
wheel==0.42.0
